% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satbathy-package.R
\docType{package}
\name{satbathy-package}
\alias{satbathy}
\alias{satbathy-package}
\title{satbathy: Satellite-Derived Bathymetry with the Empirical Ratio Transform}
\description{
Tools to estimate water depth in shallow, low-turbidity coastal waters from multispectral satellite imagery (e.g. Landsat, Sentinel-2 and PlanetScope) using the empirical log-ratio transform of Stumpf, Holderied and Sinclair (2003) \doi{10.4319/lo.2003.48.1_part_2.0547}. The package provides water masking, sun-glint correction, spatial filtering, log-ratio computation, linear model fitting and prediction, iterative estimation of the maximum reliable depth, turbidity indexing for image selection, and accuracy assessment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jfmas/satbathy}
  \item \url{https://jfmas.github.io/satbathy/}
  \item Report bugs at \url{https://github.com/jfmas/satbathy/issues}
}

}
\author{
\strong{Maintainer}: Jean François Mas \email{jfmas@ciga.unam.mx}

Authors:
\itemize{
  \item Jean François Mas \email{jfmas@ciga.unam.mx}
  \item Azucena Pérez Vega
}

}
\keyword{internal}
