% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidity.R
\name{turbidity_index}
\alias{turbidity_index}
\title{Turbidity index (Frohn and Autrey 2009)}
\usage{
turbidity_index(x, blue = "blue", green = "green", red = "red")
}
\arguments{
\item{x}{A multi-band \link[terra:SpatRaster]{terra::SpatRaster} of surface reflectance.}

\item{blue, green, red}{Band identifiers (layer names or integer indices)
for the blue, green and red bands of \code{x}.}
}
\value{
A single-layer \link[terra:SpatRaster]{terra::SpatRaster} named \code{"turbidity"}.
}
\description{
Computes the band-ratio turbidity index of Frohn and Autrey (2009),
defined as \code{(green + red) / blue}. Higher values indicate more turbid
water. It is useful for selecting the least turbid image dates before
deriving bathymetry, because the empirical ratio transform performs best
in clear water.
}
\examples{
img <- terra::rast(system.file("extdata", "example_sr.tif",
                               package = "satbathy"))
turb <- turbidity_index(img, blue = "blue", green = "green", red = "red")
turb

}
\references{
Frohn RC, Autrey BC (2009). Water quality assessment in the Ohio River
using new indices for turbidity and chlorophyll-a with Landsat-7 imagery.
Draft Internal Report, U.S. Environmental Protection Agency.
}
\seealso{
\code{\link[=ndwi]{ndwi()}}, \code{\link[=log_ratio]{log_ratio()}}
}
